move host and port determine logic from ServiceConfig to ProtocolConfig
